Changed previous removed installation and reinstalled a new version
Before the reinstallation, I encountered an error "npm error Could not
determine executable to run". After searching the internet, I found a
possible solution but it only describes ensuring that the 'tailwind'
verison is compatible. So, I tried to reinstall all the npm modules
first. This still gave me the error. So, I decided to remove
"node_modules/" and "dist/" directories and also removed package.json
and package-lock.json, the reinstalled. This solved my problem.